+++ src/components/About.jsx
@@ -1,8 +1,8 @@
 import { RichText } from "../components";
 
 const About = ({ data }) => {
-  console.log("data", data);
   return <RichText texts={data} />;
 };
 
 export default About;
+
