+++ src/components/Accordion.jsx
@@ -1,10 +1,6 @@
 import { useState, useRef } from "react";
-import {
-  AccordionSection,
-  AccordionBtn,
-  AccordionContent,
-  Chevron,
-} from "@/styles/Accordion";
+import styled from "styled-components";
+import { HiOutlineChevronDown } from "react-icons/hi";
 
 const Accordion = (props) => {
   const [active, setActive] = useState("");
@@ -37,3 +33,39 @@ const Accordion = (props) => {
   );
 };
 export default Accordion;
+
+const AccordionSection = styled.div`
+  display: flex;
+  flex-direction: column;
+  font-size: 1rem;
+  height: 100%;
+`;
+
+const AccordionBtn = styled.button`
+  background-color: black;
+  color: white;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  border: none;
+  outline: none;
+  transition: all 0.6s ease;
+  padding: 0px;
+  font-size: 20px;
+  display: flex;
+  justify-content: space-between;
+  padding: 10px 0px;
+`;
+
+const Chevron = styled(HiOutlineChevronDown)`
+  transition: transform 0.6s ease;
+  transform: ${(props) => (props.active ? "rotate(180deg)" : "none")};
+`;
+
+const AccordionContent = styled.div`
+  background-color: black;
+  color: white;
+  overflow: hidden;
+  transition: max-height 0.6s ease;
+  padding: ${(props) => (props.active ? "0rem 0rem 10rem 0rem" : "0rem")};
+`;
