+++ src/components/Awards.jsx
@@ -1,4 +1,4 @@
-import { Container, Year, Title, Prize } from "../styles/Awards";
+import styled from "styled-components";
 
 const Awards = ({ data }) => {
   const getAwards = data.map((item) => (
@@ -15,3 +15,21 @@ const Awards = ({ data }) => {
   return <div>{getAwards}</div>;
 };
 export default Awards;
+
+const Container = styled.div`
+  display: grid;
+  grid-template-columns: repeat(10, 1fr);
+  padding-bottom: 20px;
+`;
+
+const Year = styled.p`
+  grid-column: 1/3;
+`;
+
+const Title = styled.div`
+  grid-column: 3/8;
+`;
+
+const Prize = styled.p`
+  grid-column: 9/11;
+`;
