+++ src/components/Exhibitions.jsx
@@ -1,4 +1,5 @@
-import { Container, Year, Title, Country } from "@/styles/Exhibitions";
+import styled from "styled-components";
+import { device } from "../styles/device";
 
 const Exhibitions = ({ data }) => {
   const getExhibitions = data.map((item) => (
@@ -16,3 +17,28 @@ const Exhibitions = ({ data }) => {
 };
 
 export default Exhibitions;
+
+const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding: 20px 0px;
+
+  @media ${device.laptop} {
+    display: grid;
+    grid-template-columns: repeat(10, 1fr);
+    padding: 0px 0px 20px 0px;
+    font-size: 1.1rem;
+  }
+`;
+
+const Year = styled.p`
+  grid-column: 1/3;
+`;
+
+const Title = styled.div`
+  grid-column: 3/8;
+`;
+
+const Country = styled.p`
+  grid-column: 9/11;
+`;
