--- src/styles/ActiveLink.js
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const ActiveLinkStyles = styled.a`
-  color: ${(props) => (props.pathname === props.href ? "white" : "white")};
-  font-weight: ${(props) => (props.pathname === props.href ? "bold" : "regular")};
-`;
