--- src/styles/Credits.js
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const CreditText = styled.p`
-  padding: 0px 0px 10px 0px;
-`;
