--- src/styles/Home.js
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const HeroContainer = styled.div`
-  height: 100%;
-  width: 100%;
-`;
