--- src/styles/Loading.js
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const LoadingWrapper = styled.div`
-  padding-top: 20rem;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
