--- src/styles/Tags.js
@@ -1,17 +1,0 @@
-import styled from "styled-components";
-import { device } from "../styles/device";
-
-export const TagName = styled.div`
-  margin-right: 8px;
-
-  @media ${device.laptop} {
-    display: flex;
-    flex-direction: row;
-  }
-`;
-
-export const TagLink = styled.div`
-  margin-right: 8px;
-  cursor: pointer;
-  text-decoration: underline;
-`;
